move back to single model
we are going to unmarshall all old schema versions as the new schema since we maintain backwards compatibility
