+++ src/main/java/edu/ucsb/nceas/mdqengine/model/Result.java
@@ -34,7 +34,7 @@ public class Result {
 		return output;
 	}
 
-	public void setOutput(Output... output) {
+	public void setOutput(Output output) {
 		this.output = Arrays.asList(output);
 	}
