+++ src/main/java/edu/ucsb/nceas/mdqengine/model/package-info.java
@@ -1,3 +1,3 @@
-@XmlSchema(namespace="https://nceas.ucsb.edu/mdqe/v1")
+@XmlSchema(namespace = "https://nceas.ucsb.edu/mdqe/v1.2")
 package edu.ucsb.nceas.mdqengine.model;
 import javax.xml.bind.annotation.XmlSchema;
