+++ src/test/java/edu/ucsb/nceas/mdqengine/store/MDQStoreTest.java
@@ -4,6 +4,11 @@ import edu.ucsb.nceas.mdqengine.MDQEngine;
 import edu.ucsb.nceas.mdqengine.exception.MetadigException;
 import edu.ucsb.nceas.mdqengine.exception.MetadigStoreException;
 import edu.ucsb.nceas.mdqengine.model.*;
+import edu.ucsb.nceas.mdqengine.model.Check;
+import edu.ucsb.nceas.mdqengine.model.Result;
+import edu.ucsb.nceas.mdqengine.model.Run;
+import edu.ucsb.nceas.mdqengine.model.Suite;
+
 import org.dataone.service.types.v2.SystemMetadata;
 import org.apache.commons.configuration2.ex.ConfigurationException;
 
@@ -160,7 +165,7 @@ public class MDQStoreTest {
 			SystemMetadata sysMeta = null;
 			
 			Suite suite = SuiteFactory.getMockSuite();
-			Check checkRef = CheckV2.newCheck();
+			Check checkRef = new Check();
 			String checkId = "check.1.1";
 			checkRef.setId(checkId );
 			suite.getCheck().add(checkRef);
